trt-2163: move ExtractImageFromReleasePayload to payload package
